+++ vite.config.ts
@@ -1,0 +1,29 @@
+import { resolve } from 'path';
+import { defineConfig } from 'vite';
+import dts from 'vite-plugin-dts';
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [
+    dts({
+      insertTypesEntry: true,
+    }),
+  ],
+  build: {
+    lib: {
+      entry: resolve(__dirname, 'lib/index.ts'),
+      name: 'axios',
+      fileName: 'axios',
+    },
+    rollupOptions: {
+      // make sure to externalize deps that shouldn't be bundled into your library
+      external: ['axios'],
+      output: {
+        // Provide global variables to use in the UMD build for externalized deps
+        globals: {
+          axios: 'axios',
+        },
+      },
+    },
+  },
+});